refactor: clean up unnecessary changes
