+++ webpack.dev.config.js
@@ -14,9 +14,4 @@ module.exports = createConfig('webpack-dev', {
       '@edx/frontend-platform': path.resolve(__dirname, 'src'),
     },
   },
-  devServer: {
-    static: {
-      directory: path.resolve(__dirname, 'example/dist'),
-    },
-  },
 });
